Criado status task progress task

+++ src/domain/dto/auxiliar/progressTask/progressTask.go
@@ -1,6 +1,6 @@
 package progressTaskDto
 
 type ProgressTask struct {
-	TaskId    int
-	ProgresId int
+	TaskId    int64
+	ProgresId int64
 }
